requirements.txt
streamlit
fastapi
sentence-transformers
numpy
faiss-cpu
python-docx
PyMuPDF
google-generativeai
python-dotenv
beautifulsoup4
requests
pandas
openpyxl